Complete create, update, delete post

+++ src/routes/posts/posts.controller.ts
@@ -2,7 +2,7 @@ import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common'
 import { AuthType, ConditionGuard } from 'src/shared/constants/auth.constant'
 import { ActiveUser } from 'src/shared/decorators/active-user.decorator'
 import { Auth } from 'src/shared/decorators/auth.decorator'
-import { GetPostItemDTO } from './post.dto'
+import { CreatePostBodyDTO, GetPostItemDTO, UpdatePostBodyDTO } from './post.dto'
 import { PostsService } from './posts.service'
 
 @Controller('posts')
@@ -17,22 +17,24 @@ export class PostsController {
 
   @Post()
   @Auth([AuthType.Bearer])
-  createPost(@Body() body: any, @ActiveUser('userId') userId: number) {
-    return this.postService.createPost(userId, body)
+  async createPost(@Body() body: CreatePostBodyDTO, @ActiveUser('userId') userId: number) {
+    return new GetPostItemDTO(await this.postService.createPost(userId, body))
   }
 
   @Get(':id')
-  getPost(@Param('id') id: string) {
-    return this.postService.getPost(id)
+  async getPost(@Param('id') id: string) {
+    return new GetPostItemDTO(await this.postService.getPost(Number(id)))
   }
 
-  @Put('id')
-  updatePost(@Param('id') id: string, @Body() body: any) {
-    return this.postService.updatePost(id, body)
+  @Put(':id')
+  @Auth([AuthType.Bearer])
+  async updatePost(@Param('id') id: string, @ActiveUser('userId') userId: number, @Body() body: UpdatePostBodyDTO) {
+    return new GetPostItemDTO(await this.postService.updatePost({ postId: Number(id), userId, body }))
   }
 
-  @Delete('id')
-  deletePost(@Param('id') id: string) {
-    return this.postService.deletePost(id)
+  @Delete(':id')
+  @Auth([AuthType.Bearer])
+  deletePost(@Param('id') id: string, @ActiveUser('userId') userId: number): Promise<boolean> {
+    return this.postService.deletePost({ postId: Number(id), userId })
   }
 }
